third_party/WebKit/LayoutTests/media/before-load-member-access.html

<!DOCTYPE HTML>
<title>Test that accessing member of a non loaded video works.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
test(function() {
    var video = document.querySelector("video");
    assert_equals(video.played.length, 0);
    assert_equals(video.buffered.length, 0);
    assert_equals(video.seekable.length, 0);
    assert_equals(video.volume, 1);
});
</script>
